+++ lib/ripper_ruby_parser/parser.rb
@@ -12,6 +12,8 @@ module RipperRubyParser
       processor = SexpProcessor.new(filename: filename)
       result = processor.process exp
 
+      result = result[1] if result.sexp_type == :begin
+
       if result.sexp_type == :void_stmt
         nil
       else