+++ lib/ripper_ruby_parser/sexp_handlers/methods.rb
@@ -81,8 +81,9 @@ module RipperRubyParser
         when 0
           [s(:nil)]
         else
+          block = unwrap_begin(block)
           if block.sexp_type == :block
-            block.sexp_body
+            block.sexp_body.map { |it| unwrap_begin(it) }
           else
             [block]
           end