+++ lib/ripper_ruby_parser/sexp_processor.rb
@@ -79,7 +79,7 @@ module RipperRubyParser
           first.shift
           s(:block, *first, *statements)
         else
-          s(:block, first, *statements)
+          s(:block, unwrap_begin(first), *map_unwrap_begin_list(statements))
         end
       end
     end